Segue to pass values to dashboard

+++ calCounter/Model/BMRCalc.swift
@@ -15,7 +15,7 @@ class BMRCalc{
     var ageInYears: Int = 0
     var gender:Int = 0
     
-    func calculateBMR() -> Int {
+    func calculateBMR() -> Double {
         var bmr = 0.0
         var weightCalc = 0.0
         var heightCalc = 0.0
@@ -35,6 +35,6 @@ class BMRCalc{
             
             bmr = 88.362 + weightCalc + heightCalc - ageCalc
         }
-        return Int(bmr)
+        return Double(bmr)
     }
 }
